Torch.Internal.Class

Documentation

class Castable a b where

Methods

cast :: a -> (b -> IO r) -> IO r

uncast :: b -> (a -> IO r) -> IO r

class CppObject a where

Methods

fromPtr :: Ptr a -> IO (ForeignPtr a)

class CppTuple2 m where

Minimal complete definition

get0, get1

Associated Types

type A m

type B m

Methods

get0 :: m -> IO (A m)

get1 :: m -> IO (B m)

makeTuple2 :: (A m, B m) -> IO m

class CppTuple2 m => CppTuple3 m where

Minimal complete definition

get2

Associated Types

type C m

Methods

get2 :: m -> IO (C m)

makeTuple3 :: (A m, B m, C m) -> IO m

class CppTuple3 m => CppTuple4 m where

Minimal complete definition

get3

Associated Types

type D m

Methods

get3 :: m -> IO (D m)

makeTuple4 :: (A m, B m, C m, D m) -> IO m

class CppTuple4 m => CppTuple5 m where

Minimal complete definition

get4

Associated Types

type E m

Methods

get4 :: m -> IO (E m)

makeTuple5 :: (A m, B m, C m, D m, E m) -> IO m

class CppTuple5 m => CppTuple6 m where

Minimal complete definition

get5

Associated Types

type F m

Methods

get5 :: m -> IO (F m)

makeTuple6 :: (A m, B m, C m, D m, E m, F m) -> IO m