Torch.Internal.GC

Documentation

c_showWeakPtrList :: CInt -> IO ()

mallocTrim :: CInt -> IO ()

dumpLibtorchObjects

Arguments

:: Int
age

-> IO ()
output

Returns all objects of libtorch. Each time it is called, the age of the object increases by one. Dumps objects that are greater than or equal to the argument of age.

prettyException :: IO a -> IO a

retryWithGC' :: Int -> IO a -> IO a

retryWithGC :: IO a -> IO a

checkOSMemoryWithGC :: IO ()

monitorMemory :: IO () -> IO ()